applications/GL-Agent-Factory/cli/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "greenlang-agent-factory-cli"
version = "0.1.0"
description = "CLI for the GreenLang Agent Factory - Generate, validate, test, and publish AI agents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "GreenLang Team", email = "team@greenlang.io"}
]
keywords = ["ai", "agents", "cli", "greenlang", "factory"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "jinja2>=3.1.2",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
]

[project.urls]
Homepage = "https://github.com/greenlang/agent-factory"
Documentation = "https://docs.greenlang.io/agent-factory"
Repository = "https://github.com/greenlang/agent-factory"
Issues = "https://github.com/greenlang/agent-factory/issues"

[project.scripts]
gl = "cli.main:cli_main"

[tool.setuptools]
packages = ["cli", "cli.commands", "cli.utils"]

[tool.setuptools.package-data]
cli = ["templates/*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (black handles this)
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=cli",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
